Add gitlab ci definition to automatically compile
This commit adds a gitlab ci configuration that automatically builds the
pdf based on the latex project.

This ci is mainly used to validate the pushed latex code and error if
the code could not be compiled.

It is also useful as a backup for loosing your local developer
environment or downloading your last successful build from the web.